+++ R/NAMESPACE_doby.R
@@ -2,7 +2,8 @@
 #' @importFrom MASS ginv Null
 #'
 #' @importFrom Matrix rowMeans colMeans rowSums colSums
-#'
+#' @importFrom stats model.response model.weights rstandard rstudent residuals
+#' 
 #' @importClassesFrom Matrix Matrix
 #' @importFrom Matrix Matrix sparseMatrix
 #' @importMethodsFrom Matrix t isSymmetric "%*%" solve diag
@@ -10,7 +11,7 @@
 #' @importFrom stats as.formula coef delete.response family fitted
 #'     formula ftable lm logLik median model.frame model.matrix pchisq pnorm
 #'     pt qnorm qt residuals sd summary.lm terms vcov aggregate
-#'     printCoefmat confint predict resid rstandard step update
+#'     printCoefmat confint predict resid rstandard step update sigma
 #'     
 #' @importFrom modelr rmse