+++ R/by_splitBy.R
@@ -8,11 +8,13 @@
 #' 
 ###########################################################################
 #'
-#' @param formula Variables to split data frame by, as `as.quoted`
-#'     variables, a formula or character vector.
+#' @param formula A character vector or a right hand sided formula.
+#' @param ... A character vector or, right hand sided formula or a
+#'     listing of variables.
 #' @param data A data frame
-#' @param drop Logical indicating if levels that do not occur should
-#'     be dropped. Deprecated; levels that do not occur are ignored.
+#' @param omit Logical Should variables split on be omitted from the
+#'     output. Defaults to TRUE.
+#' 
 #' @param x An object.
 #' @param n A single integer.  If positive or zero, size for the
 #'     resulting object: number of elements for a vector (including
@@ -20,8 +22,6 @@
 #'     function.  If negative, all but the "n" last/first number of
 #'     elements of "x".
 #'
-#' @param ... Arguments to be passed to or from other methods.
-#' 
 #' @return A list of dataframes.
 #' 
 #' @author Søren Højsgaard, \email{sorenh@@math.aau.dk}
@@ -30,35 +30,131 @@
 #'     \code{\link{transformBy}}, \code{\link{transform_by}} 
 #' @keywords utilities
 #' @examples
+#'
+#' split_by(CO2, ~Treatment+Type)
+#' split_by(CO2, Treatment, Type)
+#' split_by(CO2, c("Treatment", "Type"))
+#' split_by(CO2, Treatment)
+#' x <- CO2 |> split_by(~Treatment)
+#' head(x, 3)
+#' tail(x, 3)
 #' 
-#' data(dietox, package="doBy")
-#' splitBy(formula = ~Evit + Cu, data = dietox)
-#' splitBy(formula = c("Evit", "Cu"), data = dietox)
-#' 
-#' splitBy(~Treatment + Type, data=CO2)
+#' x <- CO2 |> split_by(~Treatment, omit=FALSE)
+#' head(x, 3)
+#' tail(x, 3)
+#'
+#' ## The "Old" interface
+#' splitBy(~Treatment+Type, CO2)
+#' splitBy(~Treatment+Type, data=CO2)
 #' splitBy(c("Treatment", "Type"), data=CO2)
 #'
-#' x <- splitBy(~Treatment, data=CO2)
-#' head(x)
-#' tail(x)
+NULL
+
+split_by_worker <- function(data., rhs, omit=TRUE){ 
+
+    is.tib <- inherits(data., "tbl_df")
+    if (is.tib)
+        data. = as.data.frame(data.)
+    
+    grps <- apply(data.[, rhs, drop=FALSE], 1, paste0, collapse="|")
+
+    ## print(grps)
+    out <- split(data., grps)
+    
+    if (omit){
+        rhs.idx <- match(rhs, names(data.))
+        out <- lapply(out, function(d){
+            d[, -rhs.idx]            
+        })
+    }
+        
+    if (is.tib) {
+        out <- lapply(out, function(d) {
+            as_tibble(d)
+        })
+    }
+        
+    groupid <- unique(data.[, rhs, drop=FALSE])
+    idxvec <- split(1:nrow(data.), grps)
+    
+    attr(out, "groupid") <- groupid
+    attr(out, "idxvec")  <- idxvec
+    attr(out, "grps")    <- grps
+
+    class(out) <- c("splitByData", "list")    
+    return(out)
+}
 
 
 #' @export 
 #' @rdname by-split
-split_by <- function(data, formula, drop=TRUE){
-    arg <- list(formula=formula, data=data, drop=drop)
-    do.call(splitBy, arg)
+split_by <- function(data, ..., omit=TRUE){
+    dots <- match.call(expand.dots = FALSE)$...
+    rhs <- dots2char(dots)
+    split_by_worker(data, rhs, omit=omit)
 }
 
 #' @export 
 #' @rdname by-split
-splitBy <-function (formula, data = parent.frame(), drop=TRUE) {
+splitBy <- function(formula, data, omit=TRUE){
+    if (inherits(formula, "formula")){
+        rhs       <- all.vars(formula[[2]])
+    } else {
+        rhs <- formula
+    }
+    split_by_worker(data, rhs, omit=omit)
+}
+
+
+
+#' @export
+print.splitByData <- function(x, ...){
+#  print(attr(x,"groupid"))
+    print(cbind(listentry=names(x), attr(x,"groupid")))
+    return(invisible(x))
+}
+
+#' @importFrom utils head
+#' @export
+#' @rdname by-split
+head.splitByData  <- function(x, n=6L, ...){
+    lapply(x, head, n=n, ...)
+}
+
+#' @importFrom utils tail
+#' @export
+#' @rdname by-split
+tail.splitByData  <- function(x, n=6L, ...){
+    lapply(x, tail, n=n, ...)
+}
+
+
+
+
+
+
+
+
+
+
+#' @export 
+#' @rdname by-split
+#' @param drop Obsolete
+split_by.legacy <- function(data, formula, drop=TRUE){
+    arg <- list(formula=formula, data=data, drop=drop)
+    do.call(splitBy.legacy, arg)
+}
+
+#' @export 
+#' @rdname by-split
+splitBy.legacy <-function (formula, data = parent.frame(), drop=TRUE) {
                                         #, return.matrix=FALSE){
 
     if (!inherits(data, "tbl_df")) is.tib = FALSE
     else {is.tib = TRUE; data = as.data.frame(data)}
     
     data.var  <- names(data)
+    
     if (!(inherits(formula, c("formula", "character"))))        
         stop("'formula' must be a right hand sided formula or a character vector")
     
@@ -78,13 +174,16 @@ splitBy <-function (formula, data = parent.frame(), drop=TRUE) {
     fac.var   <- data.var[ !num.idx ]
     
     rhs.fac   <- intersect( rhs.var, data.var )
+
     if ("." %in% rhs.var){
         ## need all factors not mentioned elsewhere as grouping factors
         rhs.fac <- union( fac.var, rhs.fac)
     }
+
+    
     ## str(list(rhs.var=rhs.var, rhs.fac=rhs.fac, fac.var=fac.var))
     
-    rh.trivial <- length( rhs.var ) == 0 #; cat(sprintf("rh.trivial=%d\n", rh.trivial))
+    rh.trivial <- length( rhs.var ) == 0 
     
     ## FIXME rhs.fac, rhs.var -- clean up!!!
     ## Use: data, rhs.fac, rh.trivial
@@ -101,8 +200,9 @@ splitBy <-function (formula, data = parent.frame(), drop=TRUE) {
     out_list <- vector("list", length(unique.grps))
 
     names(out_list) <- unique.grps
+
     for (ii in 1:length(unique.grps)){
-        dd <- data[unique.grps[ ii ] == grps,,drop=FALSE]
+        dd <- data[unique.grps[ ii ] == grps, ,drop=FALSE]
         if (drop && length(factor.columns)>0){
             for (jj in 1:length(factor.columns)){
                 dd[, factor.columns[jj]] <- factor(dd[, factor.columns[jj]])
@@ -129,26 +229,4 @@ splitBy <-function (formula, data = parent.frame(), drop=TRUE) {
     out_list
 }
 
-
-
-#' @export
-print.splitByData <- function(x, ...){
-#  print(attr(x,"groupid"))
-    print(cbind(listentry=names(x), attr(x,"groupid")))
-    return(invisible(x))
-}
-
-#' @importFrom utils head
-#' @export
-#' @rdname by-split
-head.splitByData  <- function(x, n=6L, ...){
-    lapply(x, head, n=n, ...)
-}
-
-#' @importFrom utils tail
-#' @export
-#' @rdname by-split
-tail.splitByData  <- function(x, n=6L, ...){
-    lapply(x, tail, n=n, ...)
-}