+++ R/plot_functions.r
@@ -2,10 +2,11 @@
 #'
 #' @param lm_fit An object of class 'lm'
 #' @param format The format of the plot (or a list of plots if format is "list")
+#' @param global_aes Currently no effect.
 #'
 #' @examples
-#'
-#' m1 <- lm(speed ~ dist, data=cars)
+#' data(income)
+#' m1 <- lm(inc ~ race + educ, data=income)
 #' plot_lm(m1)
 #' plot_lm(m1, "2x2")
 #' plot_lm(m1, "1x4")
@@ -13,28 +14,38 @@
 #' plot_lm(m1, "list")
 #' 
 #' @export
-plot_lm <- function(lm_fit, format="2x2") {
-
+plot_lm <- function(lm_fit, format="2x2", global_aes=NULL) {
+    
     if(!inherits(lm_fit, "lm"))
         stop("'lm_fit' must inherit form 'lm'\n")
 
     format <- match.arg(format, c("2x2", "1x4", "4x1", "list"))
     
-    dd <- data.frame(fitted_values          = predict(lm_fit),
-                   ## residuals           = resid(lm_fit),
-                   observed_values        = predict(lm_fit)+resid(lm_fit),
-                   standardized_residuals = rstandard(lm_fit))
-    
-    pl1 <- ggplot(dd, aes(x = .data$fitted_values, y = .data$standardized_residuals)) +
+    dd <- data.frame(
+        fitted_values          = predict(lm_fit),
+        resp        = predict(lm_fit)+resid(lm_fit),
+        std_res = rstandard(lm_fit))
+
+
+    pl1 <- ggplot(dd,
+                  aes(x = .data$fitted_values,
+                      y = .data$std_res)) +
         geom_point() + geom_hline(yintercept=c(0, -1.96, 1.95))
     
-    pl2 <- ggplot(dd, aes(x=.data$fitted_values, y=.data$observed_values)) +
+    pl2 <- ggplot(dd,
+                  aes(x=.data$fitted_values,
+                      y=.data$resp)) +
         geom_point() +  geom_smooth(method="lm", formula=y ~ x, se=FALSE)
     
-    pl3 <- ggplot(dd, aes(x=.data$fitted_values, y=sqrt(abs(.data$standardized_residuals)))) + geom_point()
+    pl3 <- ggplot(dd,
+                  aes(x=.data$fitted_values,
+                      y=sqrt(abs(.data$std_res)))) +
+        geom_point()
 
-    pl4 <- ggplot(dd, aes(sample = .data$standardized_residuals)) + stat_qq() + stat_qq_line() +
-        labs(y="standardized_residuals", x = "theoretical quantiles")
+    pl4 <- ggplot(dd,
+                  aes(sample = .data$std_res)) +
+        stat_qq() + stat_qq_line() +
+        labs(y="std_res", x = "theoretical quantiles")
 
     switch(format,
            "2x2"={ cowplot::plot_grid(pl1, pl2, pl3, pl4, nrow=2) },
@@ -46,6 +57,24 @@ plot_lm <- function(lm_fit, format="2x2") {
 }
 
 
+    
+    ## aes_ <- c(list(x = predict(lm_fit),
+    ##                y = rstandard(lm_fit)),
+    ##           global_aes)
+
+    ## print(aes_)
+    
+    ## aes_ <- ggpubr_create_aes(aes_)
+
+    ## print(aes_)
+
+    ## pl1 <- ggplot(dd, aes_) +
+        ## geom_point() + geom_hline(yintercept=c(0, -1.96, 1.95))
+    
+    ## pl1 <- ggplot(dd,
+    ##               ggpubr_create_aes(list(x = .data$fitted_values,
+    ##                   y = .data$standardized_residuals))) +
+    ##     geom_point() + geom_hline(yintercept=c(0, -1.96, 1.95))
 
 
 
@@ -89,7 +118,11 @@ interaction_plot <- function(.data, .formula, interval="conf.int"){
     rr <- sym(lhs)
     s1 <- sym(rhs[1])
     s2 <- sym(rhs[2])
+    ## s2 <- sym(rhs[-1])
+
+    # lapply(rhs, sym)
     
+    ## str(list(rr, s1, s2))
     ## If lhs is transformed
     resp <- eval(.formula[[2]], .data)
     .data$RESP <- resp  ## KLUDGY
@@ -123,7 +156,12 @@ interaction_plot <- function(.data, .formula, interval="conf.int"){
                    geom_point(data = tmp, aes(y = .data$val)) +
                    geom_line(data = tmp, aes(y = .data$val, group = !!sym(s2))) + 
                    geom_errorbar(aes(ymin=.data$lwr, ymax=.data$upr), data=tmp,
-                                 width=.4, position=position_dodge(0.1))
+                                 width=.4, position=position_dodge(0.1)) +
+                   geom_point(aes(x = factor(!!sym(s1)), y = !!sym(rr2),
+                                       colour = !!sym(s2)), data=.data)
+               
+
+               
            },
            "none" = {
                pp <- ggplot(tmp, aes(x = factor(!!sym(s1)), y = .data$val,